Game UI Skeleton
Created a UI skeleton.

Includes canvas with a Question Panel and UI buttons.
The canvas was created using the MaterialsUI plugin. This allows for a
ripple effect and several other animations to be performed on the UI. It
works similarly to a regular Unity UI Canvas.

Imported a sample rigged model and gave two animations. Entry and Idle.
When opening the scene, the entry animation is played and then it loops
the Idle post.

The goal of this build is to present how the overall structure of the
game scenes will appear.