Question Functionality Added
The scene now generates one of a number of questions when the scene is
loaded. Each question has its own set of possible answers.

When an answer is selected, the console will say CORRECT or WRONG
depending on which answer was selected.

The scene will then play the QuestionOut animation and wait a specified
number of seconds before moving on.

At the moment, when the question leaves, the scene will be reloaded and
will then load a new random question.

The same question cannot be loaded twice, so if the user answered a
question already, it will not be chosen randomly again, unless all
questions have been answered at least once. The game will then reset the
list of answered questions.

These functions are all controlled using the new GameManager game object
and the GameManager.cs and Question.cs scripts which have been newly
added to the project assets.