Question and Score Adjustments
Populated the Question Array with all questions relevant to Submarine
environment.

The Timer will wait three seconds after the scene is loaded (so that the
question intro animation plays) and will then start counting down to 0.

If the Timer reaches 0, a "TIME UP" warning will appear on screen, the
buttons will deactivate and the next question scene will load.

If the user presses a button, the Timer will stop and depending on
whether the question is correct or not, the Debug.Log will display
CORRECT or WRONG.

At the moment, all the correct answers are set to the Ans1 button.